requirements.txt
time
os
PyQt